% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{find_root_ids}
\alias{find_root_ids}
\title{find_root_ids}
\usage{
find_root_ids(selected_pathways)
}
\arguments{
\item{selected_pathways}{A vecor of GO and/or REACTOME pathways IDs.}
}
\value{
A list of GO sub-root or REACTOME root ids for provided pathways.
If a certain pathway has morn than one GO sub-roots or REACTOME roots, they will be seperated by "#".
}
\description{
If you use the default pathway databases(GO Ontologyand REACTOME),this function allows you to extract GO sub-roots or REACTOME roots for certain pathways from GO or REACTOME
to help you better understanding thier the biological meanings
}
\examples{
find_root_ids(selected_pathways=c("GO:0005834","R-HSA-111469"))
}
