% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fisherVerif}
\alias{fisherVerif}
\title{fisherVerif}
\format{A list}
\usage{
fisherVerif
}
\description{
Results of one-sided Fisher's exact test and Benjamini-Hochberg p-value correction
}
\details{
The data was generated once and is cached to reproduce the simulation results in the vignettes,
  it was used to compare the performance of gerr and FET+FDR procedures
}
\keyword{datasets}
