% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{from_id2name}
\alias{from_id2name}
\title{from_id2name}
\usage{
from_id2name(selected_pathways)
}
\arguments{
\item{selected_pathways}{A list of GO and/or REACTOME pathways IDs. Each elmment is this list can be a single
id or multi-ids seperated "#"}
}
\value{
A list of GO sub-root or REACTOME root names for provided pathways.
}
\description{
If you use the default pathway databases(GO Ontologyand REACTOME),this function can help you to get pathways
names from pathways IDs.
}
\examples{
from_id2name((selected_pathways=list(c("GO:0032991#GO:0044425#GO:0044464"),"R-HSA-5357801")))
}
