% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullProbSimResDf}
\alias{fullProbSimResDf}
\title{fullProbSimResDf}
\format{A data.frame}
\usage{
fullProbSimResDf
}
\description{
Full results of simulation studies using the probabilistic generative model of GOIs
}
\details{
The data was generated once and is cached to reproduce the simulation results in the vignettes.
  Note that due to the stochastic nature of sampling, in some runs, no gene-set is selected to construct 
  GOI. They should be filtered out prior to downstream analysis.
}
\keyword{datasets}
