% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_pathway_matrix}
\alias{gene_pathway_matrix}
\title{Homo sapiens GO ontology and REACTOME ontology gene-pathway realtionship}
\format{Formal class 'dgCMatrix' [package "Matrix"]}
\source{
http://geneontology.org/docs/download-ontology/, https://reactome.org/download-data
A binary matrix whose columns are the pathways/gene sets from GO ontology and REATOME database and
whose rows are all the genes(represented by gene HUGO gene symbols) from  GO ontology and REATOME database.
For gene i and pathway j, the value of matrix(i,j) is 1 is gene i belonging to pathway j otherwise 0
}
\usage{
gene_pathway_matrix
}
\description{
A rds R object contains GO ontology and REACTOME ontology gene-pathway realtionship
}
\keyword{datasets}
