% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{get_steps}
\alias{get_steps}
\title{get_steps}
\usage{
get_steps(selected_pathways)
}
\arguments{
\item{selected_pathways}{A vecor of GO and/or REACTOME pathways IDs.}
}
\value{
A list contains distances from pathways to GO root or REACTOME root nodes
}
\description{
If you use the default pathway databases(GO Ontologyand REACTOME),this function allows you to extract  the distances
 from ceatain pathways to  GO roots or REACTOME roots nodes.
}
\examples{
get_steps(selected_pathways=c("GO:0005834","R-HSA-111469"))
}
