% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_go_ontology}
\alias{human_go_ontology}
\title{Homo sapiens GO ontology tree}
\format{Directed igraph format}
\source{
http://geneontology.org/docs/download-ontology/
}
\usage{
readRDS(system.file("data", "human_go_ontology.rds", package = "GENEMABR"))
}
\description{
A rds R object contains GO ontology relationships (tree structure)
}
\details{
The igraph format tree was constructed by using data from http://geneontology.org/docs/download-ontology/ (May 2108)
It has three root notes representing Molecular Function,Cellular Component and Biological Process (http://geneontology.org/docs/ontology-documentation/)
}
\keyword{datasets}
