% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_go_roots}
\alias{human_go_roots}
\title{human_go_roots}
\format{A vector of GO ontology root notes (ID)}
\source{
http://geneontology.org/docs/download-ontology/
}
\usage{
readRDS(system.file("data", "human_go_roots.rds", package = "GENEMABR"))
}
\description{
A rds R object contains GO ontology tree root nodes
}
\details{
You can view tree stuctor of GO ontology at  https://www.ebi.ac.uk/QuickGO/
Thre are three roots notes in GO ontology tree: GO:0008150 (biological_process), GO:0003674(molecular_function), GO:0005575(cellular_component)
}
\keyword{datasets}
