% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_go_sub_roots}
\alias{human_go_sub_roots}
\title{human_go_sub_roots}
\format{A list of three elements contains GO ontology sub-root notes (ID)/the children of three root notes}
\source{
http://geneontology.org/docs/download-ontology/
}
\usage{
readRDS(system.file("data", "human_go_sub_roots.rds", package = "GENEMABR"))
}
\description{
A rds R object contains GO ontology tree sub-root nodes (The children of root nodes).
}
\details{
You can view tree stuctor of GO ontology at  https://www.ebi.ac.uk/QuickGO/
}
\keyword{datasets}
