% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_reactome_ontology}
\alias{human_reactome_ontology}
\title{Homo sapiens REACTOME ontology tree}
\format{Directed igraph format}
\source{
https://reactome.org/download-data
}
\usage{
readRDS(system.file("data", "human_reactome_ontology.rds", package = "GENEMABR"))
}
\description{
A rds R object contains Reactome ontology relationships (tree structure)
}
\details{
The igraph format tree was constructed by using  data from https://reactome.org/download-data (May 2108)
It has several root nodes representing REACTOME pathway categories (https://reactome.org/PathwayBrowser/)
}
\keyword{datasets}
