% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_reactome_roots}
\alias{human_reactome_roots}
\title{human_reactome_roots}
\format{A vector of REACTOME root notes (ID)}
\source{
https://reactome.org/download-data
}
\usage{
human_reactome_roots
}
\description{
A rds R object contains REACTOME tree root nodes
}
\details{
You can view tree stuctor of REACTOME at https://reactome.org/PathwayBrowser/
}
\keyword{datasets}
