% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{mydata}
\alias{mydata}
\title{Helper function to load file}
\usage{
mydata(objName, package)
}
\arguments{
\item{objName}{Object name to be loaded}

\item{package}{Package name to be loaded}
}
\value{
The object
This function is temporarily set up to help the transition from .rds to .RData file
}
\description{
Helper function to load file
}
