% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pretrained_gaussian_lambda}
\alias{pretrained_gaussian_lambda}
\title{Pretrained lambda values for Gaussian glmnet}
\format{A vector of numeric values}
\usage{
pretrained_gaussian_lambda
}
\description{
Pretrained lambda values for Gaussian glmnet
}
\keyword{datasets}
