% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_selected_pathways_funcs.R
\name{regression_selected_pathways}
\alias{regression_selected_pathways}
\title{regression_selected_pathways}
\usage{
regression_selected_pathways(gene_input, gene_pathway_matrix = NULL,
  alpha = 0.5, family = c("gaussian", "binomial"), lambda = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{gene_input}{A vecor of genes to be annotated. It should have same ID types(Ensembl ID, HUGO gene symbol)
as the genes in \emph{gene_pathway_matrix}.}

\item{gene_pathway_matrix}{A binary background matrix whose columns are the pathways/gene sets and
whose rows are all the genes from pathways/gene sets . It could be in sparse matrix format ((inherit from class
"sparseMatrix" as in package Matrix) to save memory.
For gene i and pathway j, the value of matrix(i,j) is 1 is gene i belonging to pathway j otherwise 0.
Users could leave it as default value then  it will use pre-collected gene_pathway_matrix from GO Ontology and
REACTOME databaase.
Otherwise, they could use their own customized gene_pathway_matrix}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0~\leq~\alpha~\leq~1}. The penalty is defined as
\eqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1}.
alpha=1 is the lasso penalty, and alpha=0 the ridge penalty. Default value: 0.5.}

\item{family}{Response type, currently \code{gaussian} and \code{binomial} are supported and the gaussian family is the default. Future extensions are likely.}

\item{lambda}{A user supplied lambda sequence, see \code{\link{glmnet}} and use with care.}

\item{verbose}{If supprese warning messesge from gerr functions. TRUE or FALSE. Default value TRUE, not suppressing warning messages.}

\item{...}{Other paramaters passed to the \code{\link{cv.glmnet}} function.}
}
\value{
A list of following elements:
\itemize{
  \item selected_pathways_names - Pathways names for selected pathways
  \item selected_pathways_coef - Regression coefficients value for selected pathways
  \item selected_pathways_fisher_pvalue - Fisher exact pvalue for selected pathways
  \item selected_pathways_num_genes - The number of genes for selected pathways in background
  \item model - the trained glmnet model
  \item x - independent variable
  \item y - dependent variable
}
}
\description{
This function allows you to extracte enriched pathways for gene module/list via regressioin (elastic net)
based method
}
\examples{
rspResults <- regression_selected_pathways(gene_input=c("TRPC4AP","CDC37",
  "TNIP1","IKBKB","NKIRAS2", "NFKBIA","TIMM50","RELB","TNFAIP3","NFKBIB",
  "HSPA1A","NFKBIE","SPAG9","NFKB2","ERLIN1","REL","TNIP2",
  "TUBB6","MAP3K8"),
 gene_pathway_matrix=NULL,lambda=NULL,alpha=0.5)
}
