% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simGenesets}
\alias{simGenesets}
\title{simGenesets}
\format{A list of gene-sets (in the form of character vectors), with gene-set names as the names of the list.}
\source{
The \href{https://github.com/igordot/msigdbr}{msigdbr} package, and 
  the \href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB: Molecular Signatures Database}
}
\usage{
simGenesets
}
\description{
A list of gene-sets used for simulation studies
}
\keyword{datasets}
