% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verifKeyRes}
\alias{verifKeyRes}
\title{verifKeyRes}
\format{A list of key results of gerr model verification}
\usage{
verifKeyRes
}
\description{
Key results of model verification
}
\details{
The data was generated once and is cached to reproduce the simulation results in the vignettes.
}
\keyword{datasets}
